test: accelerate sqlite reliability suite
